+++ src/components/buttons/button/Button.stories.ts
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/svelte-vite";
 import Button from "./Button.svelte";
+import { fn } from '@storybook/test';
 import { iconNames } from "../../../icons/types/IconName";
 import {
   buttonPriorities,
@@ -62,6 +63,7 @@ export const Primary: Story = {
   args: {
     priority: "primary",
     label: "Primary Button",
+    onClick: fn()
   },
 };
 
@@ -71,6 +73,7 @@ export const PrimaryCompact: Story = {
     size: "compact",
     label: "Karten App",
     iconName: "location",
+    onClick: fn()
   },
 };
 
@@ -79,6 +82,7 @@ export const SecondaryLarge: Story = {
     priority: "secondary",
     size: "large",
     label: "Secondary Button",
+    onClick: fn()
   },
 };
 
@@ -86,6 +90,7 @@ export const Tertiary: Story = {
   args: {
     priority: "tertiary",
     label: "Tertiary Button",
+    onClick: fn()
   },
 };
 
@@ -94,6 +99,7 @@ export const Newsletter: Story = {
     priority: "tertiary",
     label: "Newsletter",
     iconName: "newsletter",
+    onClick: fn()
   },
 };
 
@@ -102,6 +108,7 @@ export const Print: Story = {
     priority: "primary",
     label: "Print",
     iconName: "printer",
+    onClick: fn()
   },
 };
 
@@ -111,6 +118,7 @@ export const CustomColorsExample1: Story = {
     iconName: "facebook",
     customBGColor: "#3b5998",
     customColor: "white",
+    onClick: fn()
   },
 };
 
@@ -120,5 +128,6 @@ export const CustomColorsExample2: Story = {
     iconName: "pinterest",
     customBGColor: "rgb(189, 8, 28)",
     customColor: "white",
+    onClick: fn()
   },
 };